refactor: update image directories and package file
