+++ index.html
@@ -3,7 +3,7 @@
 
 <head>
   <meta charset="UTF-8" />
-  <link rel="icon" type="image/svg+xml" href="./public/favicon.ico" />
+  <link rel="icon" type="image/svg+xml" href="src/assets/icons/favicon.ico" />
   <meta name="viewport" content="width=device-width, initial-scale=1.0" />
   <title>Beer Сatalog</title>
 </head>
